Add controller subsystem to manage player movement.
