+++ new/main.py
@@ -4,6 +4,6 @@ import ppb
 
 from shared import TITLE
 from scenes import TitleScreen, Sandbox
-from systems import ScoreSystem
+from systems import ScoreSystem, Controller
 
-ppb.run(starting_scene=TitleScreen, title=TITLE, systems=[ScoreSystem])
+ppb.run(starting_scene=TitleScreen, title=TITLE, systems=[ScoreSystem, Controller])
