+++ new/systems.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
+from dataclasses import dataclass
 
 import ppb
-from ppb import systemslib
+from ppb import keycodes, systemslib
 
 import events
 from shared import FONT
@@ -39,3 +40,38 @@ class ScoreSystem(systemslib.System):
     def on_pre_render(self, event, signal):
         for score_display in event.scene.get(kind=ScoreDisplay):
             score_display.score = self.current_score
+
+
+class Controller(systemslib.System):
+    """"""
+    # 1. A controls object: store state of significant keys (WASD)
+    # 2. Respond to keyboard and mouse events.
+    # 3. Extensions - put control object on events (update event)
+    move_vector = ppb.Vector(0, 0)
+
+    def __init__(self, engine: ppb.engine.GameEngine, **kwargs):
+        super().__init__(engine=engine, **kwargs)
+        engine.register(ppb.events.Update, self.add_controls)
+
+    def add_controls(self, event):
+        event.movement = self.move_vector
+
+    def on_key_pressed(self, event: ppb.events.KeyPressed, signal):
+        if event.key is keycodes.W:
+            self.move_vector += ppb.directions.Up
+        elif event.key is keycodes.A:
+            self.move_vector += ppb.directions.Left
+        elif event.key is keycodes.S:
+            self.move_vector += ppb.directions.Down
+        elif event.key is keycodes.D:
+            self.move_vector += ppb.directions.Right
+
+    def on_key_released(self, event: ppb.events.KeyReleased, signal):
+        if event.key is keycodes.W:
+            self.move_vector -= ppb.directions.Up
+        elif event.key is keycodes.A:
+            self.move_vector -= ppb.directions.Left
+        elif event.key is keycodes.S:
+            self.move_vector -= ppb.directions.Down
+        elif event.key is keycodes.D:
+            self.move_vector -= ppb.directions.Right
